New package o.b.d.automation for non-FC-UI-specific automation
