+++ automation/src/test/scala/org/broadinstitute/dsde/automation/api/RestException.scala
@@ -1,0 +1,3 @@
+package org.broadinstitute.dsde.automation.api
+
+case class RestException(message: String = null, cause: Throwable = null) extends Exception(message, cause)
