+++ automation/src/test/scala/org/broadinstitute/dsde/automation/config/Credentials.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.firecloud.config
+package org.broadinstitute.dsde.automation.config
 
 case class Credentials (email: String, password: String)
