+++ automation/src/test/scala/org/broadinstitute/dsde/automation/util/ExceptionHandling.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.firecloud.util
+package org.broadinstitute.dsde.automation.util
 
 import com.typesafe.scalalogging.LazyLogging
