+++ automation/src/test/scala/org/broadinstitute/dsde/automation/util/Retry.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.firecloud.util
+package org.broadinstitute.dsde.automation.util
 
 import com.typesafe.scalalogging.LazyLogging
