--- automation/src/test/scala/org/broadinstitute/dsde/firecloud/api/APIException.scala
@@ -1,4 +1,0 @@
-package org.broadinstitute.dsde.firecloud.api
-
-
-case class APIException (message: String = null, cause: Throwable = null) extends Exception(message, cause)
