+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/library/DataLibraryPage.scala
@@ -1,7 +1,8 @@
 package org.broadinstitute.dsde.firecloud.page.library
 
+import org.broadinstitute.dsde.automation.browser.PageUtil
 import org.broadinstitute.dsde.firecloud.config.Config
-import org.broadinstitute.dsde.firecloud.page.{AuthenticatedPage, PageUtil}
+import org.broadinstitute.dsde.firecloud.page.AuthenticatedPage
 import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.Page
