+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/WorkspaceSummaryPage.scala
@@ -1,10 +1,11 @@
 package org.broadinstitute.dsde.firecloud.page.workspaces
 
+import org.broadinstitute.dsde.automation.browser.PageUtil
 import org.broadinstitute.dsde.firecloud.api.WorkspaceAccessLevel
 import org.broadinstitute.dsde.firecloud.api.WorkspaceAccessLevel.WorkspaceAccessLevel
 import org.broadinstitute.dsde.firecloud.config.Config
 import org.broadinstitute.dsde.firecloud.page.components.{Collapse, Table}
-import org.broadinstitute.dsde.firecloud.page.{PageUtil, _}
+import org.broadinstitute.dsde.firecloud.page._
 import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.Page
