+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/monitor/WorkspaceMonitorPage.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsde.firecloud.page.workspaces.monitor
 
+import org.broadinstitute.dsde.automation.browser.PageUtil
 import org.broadinstitute.dsde.firecloud.config.Config
-import org.broadinstitute.dsde.firecloud.page.PageUtil
 import org.broadinstitute.dsde.firecloud.page.workspaces.WorkspacePage
 import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.Page
